LayoutTests/imported/w3c/web-platform-tests/css/css-conditional/container-queries/column-spanner-in-container-expected.txt


FAIL #spanner matching container with column-width 300px, getting column-span:all assert_equals: expected "600px" but got "300px"
PASS Reducing #multicol width means #spanner no longer gets column-span:all
FAIL Back to matching 300px and column-span:all assert_equals: expected "600px" but got "300px"
